move floodfill saga code into own module
