+++ frontend/javascripts/oxalis/model/sagas/volumetracing_saga.tsx
@@ -1,41 +1,27 @@
-import { message } from "antd";
 import { diffDiffableMaps } from "libs/diffable_map";
 import { V3 } from "libs/mjs";
-import createProgressCallback from "libs/progress_callback";
 import Toast from "libs/toast";
-import * as Utils from "libs/utils";
 import _ from "lodash";
 import memoizeOne from "memoize-one";
 import type {
   AnnotationTool,
-  BoundingBoxType,
   ContourMode,
-  LabeledVoxelsMap,
   OrthoView,
   OverwriteMode,
   Vector3,
 } from "oxalis/constants";
-import Constants, {
+import {
   AnnotationToolEnum,
   ContourModeEnum,
-  FillModeEnum,
   OrthoViews,
-  Unicode,
   OverwriteModeEnum,
 } from "oxalis/constants";
 import getSceneController from "oxalis/controller/scene_controller_provider";
 import { CONTOUR_COLOR_DELETE, CONTOUR_COLOR_NORMAL } from "oxalis/geometries/helper_geometries";
 
-import {
-  getDatasetBoundingBox,
-  getMaximumSegmentIdForLayer,
-  getMagInfo,
-} from "oxalis/model/accessors/dataset_accessor";
-import {
-  getPosition,
-  getActiveMagIndexForLayer,
-  getRotation,
-} from "oxalis/model/accessors/flycam_accessor";
+import messages from "messages";
+import { getMaximumSegmentIdForLayer } from "oxalis/model/accessors/dataset_accessor";
+import { getPosition, getRotation } from "oxalis/model/accessors/flycam_accessor";
 import {
   isBrushTool,
   isTraceTool,
@@ -56,7 +42,6 @@ import type {
   AddAdHocMeshAction,
   AddPrecomputedMeshAction,
 } from "oxalis/model/actions/annotation_actions";
-import { addUserBoundingBoxAction } from "oxalis/model/actions/annotation_actions";
 import {
   updateTemporarySettingAction,
   updateUserSettingAction,
@@ -64,9 +49,9 @@ import {
 import { setBusyBlockingInfoAction, setToolAction } from "oxalis/model/actions/ui_actions";
 import type {
   ClickSegmentAction,
-  SetActiveCellAction,
   CreateCellAction,
   DeleteSegmentDataAction,
+  SetActiveCellAction,
 } from "oxalis/model/actions/volumetracing_actions";
 import {
   finishAnnotationStrokeAction,
@@ -74,9 +59,7 @@ import {
   setSelectedSegmentsOrGroupAction,
   updateSegmentAction,
 } from "oxalis/model/actions/volumetracing_actions";
-import BoundingBox from "oxalis/model/bucket_data_handling/bounding_box";
 import { markVolumeTransactionEnd } from "oxalis/model/bucket_data_handling/bucket";
-import Dimensions from "oxalis/model/dimensions";
 import type { Saga } from "oxalis/model/sagas/effect-generators";
 import { select, take } from "oxalis/model/sagas/effect-generators";
 import listenToMinCut from "oxalis/model/sagas/min_cut_saga";
@@ -86,33 +69,26 @@ import {
   takeEveryUnlessBusy,
 } from "oxalis/model/sagas/saga_helpers";
 import {
-  deleteSegmentDataVolumeAction,
-  type UpdateAction,
-  updateSegmentGroups,
-} from "oxalis/model/sagas/update_actions";
-import {
   createSegmentVolumeAction,
+  deleteSegmentDataVolumeAction,
   deleteSegmentVolumeAction,
   removeFallbackLayer,
+  updateMappingName,
+  updateSegmentGroups,
   updateSegmentVolumeAction,
   updateUserBoundingBoxes,
   updateVolumeTracing,
-  updateMappingName,
+  type UpdateAction,
 } from "oxalis/model/sagas/update_actions";
 import type VolumeLayer from "oxalis/model/volumetracing/volumelayer";
 import { Model, api } from "oxalis/singletons";
 import type { Flycam, SegmentMap, VolumeTracing } from "oxalis/store";
-import { actionChannel, call, fork, put, takeEvery, takeLatest } from "typed-redux-saga";
-import {
-  applyLabeledVoxelMapToAllMissingMags,
-  createVolumeLayer,
-  labelWithVoxelBuffer2D,
-  type BooleanBox,
-} from "./volume/helpers";
-import maybeInterpolateSegmentationLayer from "./volume/volume_interpolation_saga";
-import messages from "messages";
-import { pushSaveQueueTransaction } from "../actions/save_actions";
 import type { ActionPattern } from "redux-saga/effects";
+import { actionChannel, call, fork, put, takeEvery, takeLatest } from "typed-redux-saga";
+import { pushSaveQueueTransaction } from "../actions/save_actions";
+import { createVolumeLayer, labelWithVoxelBuffer2D, type BooleanBox } from "./volume/helpers";
+import maybeInterpolateSegmentationLayer from "./volume/volume_interpolation_saga";
+import { floodFill } from "./volume/floodfill_saga";
 
 const OVERWRITE_EMPTY_WARNING_KEY = "OVERWRITE-EMPTY-WARNING";
 
@@ -366,212 +342,6 @@ export function* editVolumeLayerAsync(): Saga<any> {
         );
     } else {
       yield* call([Toast, Toast.close], OVERWRITE_EMPTY_WARNING_KEY);
-    }
-  }
-}
-
-function* getBoundingBoxForFloodFill(
-  position: Vector3,
-  currentViewport: OrthoView,
-): Saga<BoundingBoxType> {
-  const fillMode = yield* select((state) => state.userConfiguration.fillMode);
-  const halfBoundingBoxSizeUVW = V3.scale(Constants.FLOOD_FILL_EXTENTS[fillMode], 0.5);
-  const currentViewportBounding = {
-    min: V3.sub(position, halfBoundingBoxSizeUVW),
-    max: V3.add(position, halfBoundingBoxSizeUVW),
-  };
-
-  if (fillMode === FillModeEnum._2D) {
-    // Only use current plane
-    const thirdDimension = Dimensions.thirdDimensionForPlane(currentViewport);
-    const numberOfSlices = 1;
-    currentViewportBounding.min[thirdDimension] = position[thirdDimension];
-    currentViewportBounding.max[thirdDimension] = position[thirdDimension] + numberOfSlices;
-  }
-
-  const datasetBoundingBox = yield* select((state) => getDatasetBoundingBox(state.dataset));
-  const { min: clippedMin, max: clippedMax } = new BoundingBox(
-    currentViewportBounding,
-  ).intersectedWith(datasetBoundingBox);
-  return {
-    min: clippedMin,
-    max: clippedMax,
-  };
-}
-
-const FLOODFILL_PROGRESS_KEY = "FLOODFILL_PROGRESS_KEY";
-export function* floodFill(): Saga<void> {
-  yield* take("INITIALIZE_VOLUMETRACING");
-  const allowUpdate = yield* select((state) => state.tracing.restrictions.allowUpdate);
-
-  while (allowUpdate) {
-    const floodFillAction = yield* take("FLOOD_FILL");
-
-    if (floodFillAction.type !== "FLOOD_FILL") {
-      throw new Error("Unexpected action. Satisfy typescript.");
-    }
-
-    const { position: positionFloat, planeId } = floodFillAction;
-    const volumeTracing = yield* select(enforceActiveVolumeTracing);
-    if (volumeTracing.hasEditableMapping) {
-      const message = "Volume modification is not allowed when an editable mapping is active.";
-      Toast.error(message);
-      console.error(message);
-      continue;
-    }
-    const segmentationLayer = yield* call(
-      [Model, Model.getSegmentationTracingLayer],
-      volumeTracing.tracingId,
-    );
-    const { cube } = segmentationLayer;
-    const seedPosition = Dimensions.roundCoordinate(positionFloat);
-    const activeCellId = volumeTracing.activeCellId;
-    const dimensionIndices = Dimensions.getIndices(planeId);
-    const requestedZoomStep = yield* select((state) =>
-      getActiveMagIndexForLayer(state, segmentationLayer.name),
-    );
-    const magInfo = yield* call(getMagInfo, segmentationLayer.mags);
-    const labeledZoomStep = magInfo.getClosestExistingIndex(requestedZoomStep);
-    const additionalCoordinates = yield* select((state) => state.flycam.additionalCoordinates);
-    const oldSegmentIdAtSeed = cube.getDataValue(
-      seedPosition,
-      additionalCoordinates,
-      null,
-      labeledZoomStep,
-    );
-
-    if (activeCellId === oldSegmentIdAtSeed) {
-      Toast.warning("The clicked voxel's id is already equal to the active segment id.");
-      continue;
-    }
-
-    const busyBlockingInfo = yield* select((state) => state.uiInformation.busyBlockingInfo);
-
-    if (busyBlockingInfo.isBusy) {
-      console.warn(`Ignoring floodfill request (reason: ${busyBlockingInfo.reason || "unknown"})`);
-      continue;
-    }
-    // As the flood fill will be applied to the volume layer,
-    // the potentially existing mapping should be locked to ensure a consistent state.
-    const isModificationAllowed = yield* call(
-      requestBucketModificationInVolumeTracing,
-      volumeTracing,
-    );
-    if (!isModificationAllowed) {
-      continue;
-    }
-    yield* put(setBusyBlockingInfoAction(true, "Floodfill is being computed."));
-    const boundingBoxForFloodFill = yield* call(getBoundingBoxForFloodFill, seedPosition, planeId);
-    const progressCallback = createProgressCallback({
-      pauseDelay: 200,
-      successMessageDelay: 2000,
-      // Since only one floodfill operation can be active at any time,
-      // a hardcoded key is sufficient.
-      key: "FLOODFILL_PROGRESS_KEY",
-    });
-    yield* call(progressCallback, false, "Performing floodfill...");
-    console.time("cube.floodFill");
-    const fillMode = yield* select((state) => state.userConfiguration.fillMode);
-
-    const {
-      bucketsWithLabeledVoxelsMap: labelMasksByBucketAndW,
-      wasBoundingBoxExceeded,
-      coveredBoundingBox,
-    } = yield* call(
-      { context: cube, fn: cube.floodFill },
-      seedPosition,
-      additionalCoordinates,
-      activeCellId,
-      dimensionIndices,
-      boundingBoxForFloodFill,
-      labeledZoomStep,
-      progressCallback,
-      fillMode === FillModeEnum._3D,
-    );
-    console.timeEnd("cube.floodFill");
-    yield* call(progressCallback, false, "Finalizing floodfill...");
-    const indexSet: Set<number> = new Set();
-
-    for (const labelMaskByIndex of labelMasksByBucketAndW.values()) {
-      for (const zIndex of labelMaskByIndex.keys()) {
-        indexSet.add(zIndex);
-      }
-    }
-
-    console.time("applyLabeledVoxelMapToAllMissingMags");
-
-    for (const indexZ of indexSet) {
-      const labeledVoxelMapFromFloodFill: LabeledVoxelsMap = new Map();
-
-      for (const [bucketAddress, labelMaskByIndex] of labelMasksByBucketAndW.entries()) {
-        const map = labelMaskByIndex.get(indexZ);
-
-        if (map != null) {
-          labeledVoxelMapFromFloodFill.set(bucketAddress, map);
-        }
-      }
-
-      applyLabeledVoxelMapToAllMissingMags(
-        labeledVoxelMapFromFloodFill,
-        labeledZoomStep,
-        dimensionIndices,
-        magInfo,
-        cube,
-        activeCellId,
-        indexZ,
-        true,
-      );
-    }
-
-    yield* put(finishAnnotationStrokeAction(volumeTracing.tracingId));
-    yield* put(
-      updateSegmentAction(
-        volumeTracing.activeCellId,
-        {
-          somePosition: seedPosition,
-          someAdditionalCoordinates: additionalCoordinates || undefined,
-        },
-        volumeTracing.tracingId,
-      ),
-    );
-
-    console.timeEnd("applyLabeledVoxelMapToAllMissingMags");
-
-    if (wasBoundingBoxExceeded) {
-      yield* call(
-        progressCallback,
-        true,
-        <>
-          Floodfill is done, but terminated since the labeled volume got too large. A bounding box
-          <br />
-          that represents the labeled volume was added.{Unicode.NonBreakingSpace}
-          <a href="#" onClick={() => message.destroy(FLOODFILL_PROGRESS_KEY)}>
-            Close
-          </a>
-        </>,
-        {
-          successMessageDelay: 10000,
-        },
-      );
-      yield* put(
-        addUserBoundingBoxAction({
-          boundingBox: coveredBoundingBox,
-          name: `Limits of flood-fill (source_id=${oldSegmentIdAtSeed}, target_id=${activeCellId}, seed=${seedPosition.join(
-            ",",
-          )}, timestamp=${new Date().getTime()})`,
-          color: Utils.getRandomColor(),
-          isVisible: true,
-        }),
-      );
-    } else {
-      yield* call(progressCallback, true, "Floodfill done.");
-    }
-
-    cube.triggerPushQueue();
-    yield* put(setBusyBlockingInfoAction(false));
-
-    if (floodFillAction.callback != null) {
-      floodFillAction.callback();
     }
   }
 }
